added calendar to nutrition
