+++ package.json
@@ -10,9 +10,11 @@
     "@mui/joy": "^5.0.0-alpha.76",
     "@mui/material": "^5.12.0",
     "@mui/styled-engine-sc": "^5.12.0",
+    "@mui/x-date-pickers": "^6.2.1",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "dayjs": "^1.11.7",
     "firebase": "^9.20.0",
     "firebaseui": "^6.0.2",
     "react": "^18.2.0",
